refactor: 상품 관련 유틸 함수 분리하기
