+++ src/refactoring/utils/productUtils.ts
